requirements.txt
algoliasearch==2.6.2
algoliasearch-django==2.0.0
asgiref==3.4.1
bcrypt==3.2.0
black==22.1.0
boto3==1.21.6
botocore==1.24.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
Markdown==3.3.4
MarkupSafe==2.1.0
mypy-extensions==0.4.3
packaging==21.3
paramiko==2.9.2
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
storage==0.0.4.3
tomli==1.2.2
typing_extensions==4.0.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.14.0
whitenoise==6.0.0